Removes old empty file
